backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "google-api-python-client>=2.188.0",
    "google-auth>=2.48.0",
    "google-auth-oauthlib>=1.2.4",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.46",
    "sqlalchemy-utils>=0.42.1",
    "uvicorn>=0.40.0",
]
